requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
Pillow
pytesseract
python-multipart
pandas
scikit-learn
tqdm
opencv-python
numpy
pillow
pytesseract
easyocr
torch
torchvision
scikit-learn
fastapi
uvicorn
pytest
notebook
python-multipart
tqdm
pandas
faiss-cpu
